# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BooleanIT-Task API Documentation",
        "description": "API documentation for BooleanIT task",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "description": "Get all categories.",
                "operationId": "af8a7c6b1a539253e3246700fd1aaedc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of categories per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{categoryId}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "description": "Update a category.",
                "operationId": "88ab48a84b7dea0474959037c58c417c",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_name"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "example": "Updated Category Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "description": "Delete a category.",
                "operationId": "0e3cf4d72ecdda8b7654b96cde907751",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Category 1 deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{categoryId}/products": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get products in a category",
                "description": "Get products in a category.",
                "operationId": "e81ac206e90523d174f972630f969245",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{categoryId}/products/export": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Export category products as CSV",
                "description": "Exports all products in a given category as a downloadable CSV file.",
                "operationId": "7a6a4b35efbf683a28b2e36f35cc048b",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file download",
                        "headers": {
                            "Content-Disposition": {
                                "description": "attachment; filename=category_products.csv",
                                "schema": {
                                    "type": "string",
                                    "example": "attachment; filename=category_products.csv"
                                }
                            }
                        },
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Get all products.",
                "operationId": "2edf916735efe3678721a3da72a4a052",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of products per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Update a product.",
                "operationId": "ef3ac92306a3827ff30f98efc6f49448",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "regular_price"
                                ],
                                "properties": {
                                    "regular_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 999.99
                                    },
                                    "sale_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 799.99,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Delete a product.",
                "operationId": "93b8e5e09d76ade9578cdc1adb6366e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product 1 has been deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category model",
                "required": [
                    "category_id",
                    "category_name",
                    "status"
                ],
                "properties": {
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "deleted"
                        ],
                        "example": "active"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "description": "Product model",
                "required": [
                    "product_id",
                    "category_id",
                    "department_id",
                    "manufacturer_id",
                    "product_number",
                    "regular_price",
                    "status"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "manufacturer_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "product_number": {
                        "type": "string",
                        "example": "123456"
                    },
                    "sku": {
                        "type": "string",
                        "example": "987654",
                        "nullable": true
                    },
                    "upc": {
                        "type": "string",
                        "example": "123456789012",
                        "nullable": true
                    },
                    "regular_price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "sale_price": {
                        "type": "number",
                        "format": "float",
                        "example": 799.99,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "deleted"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "Endpoints related to categories"
        },
        {
            "name": "Products",
            "description": "Endpoints related to products"
        }
    ]
}